fix: 修复教学项目 P0 反模式（密钥/金额/JWT/CORS/上传/分页）
- 密钥：JwtSecret / SessionSecret / MoneySecret 走 config + 环境变量
  覆盖（JWT_SECRET / SESSION_SECRET / MONEY_SECRET），启动 fail-fast。
- JWT：EmailClaims 改用 PasswordDigest（bcrypt 哈希），不再带明文密码。
- 金额：Order.Money / DecryptMoney / PaymentDownReq.Money
  统一改 int64（分），strconv.FormatInt 替代 fmt.Sprintf("%f")。
- 支付鉴权：PaymentDownReq 仅保留 OrderId + Key；
  bossID/productID/num 全部从 order 取，新增已支付订单挡板与 Key 长度校验。
- 错误处理：补齐 product/payment 链路被忽略的 err。
- CORS：去掉 Allow-Origin: *，按 System.AllowedOrigins 白名单回显 Origin。
- Cookie：access/refresh cookie 增加 SameSite=Lax。
- 上传：替换 ioutil；按 MIME 校验 jpeg/png/gif/webp，限 5MB，
  权限 0644，文件名 sanitize。
- 分页：BasePage.Normalize() 限 PageNum/PageSize/LastId，
  新增 consts.MaxPageSize=100。
- 文档：新增 doc/MIGRATION.md，新增 config/locales/config.example.yaml。

+++ repository/db/model/order.go
@@ -14,5 +14,5 @@ type Order struct {
 	Num       int    // 数量
 	OrderNum  uint64 // 订单号
 	Type      uint   // 1 未支付  2 已支付
-	Money     float64
+	Money int64 // 单位：分
 }
